Rename to flutter_runtime_env
It just makes more sense. Enabling or disabling analytics is a decission
that is very buisness specific. Even in GitJournal we were using this
method + additional logic.

Plus, for example - I really don't want my own personal devices to count
towards analytics.

There isn't a huge penalty if my analytics are bit skewed, but in the
case of when to display Ads, this can be quite important.